requirements.txt
fastapi
pandas
vercel
uvicorn
dotenv
PyGithub
openai
Pillow
bs4
httpx
lxml_html_clean
requests-html==0.10.0
pyppeteer==1.0.2
websockets==10.4
langchain
langchain-openai
langchain-community
langchain-core
SpeechRecognition
pydub
playwright
